requirements.txt
# FastAPI + Web Server
fastapi
uvicorn[standard]

# Crypto & Financial Data
requests
pandas
ta
yfinance

# Firebase: Firestore + FCM
firebase-admin
pyfcm

# LangChain + OpenAI
langchain
openai
langchain-openai

# Environment Variables
python-dotenv

# Scheduler
schedule
